+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [Unreleased]
+### Changed
+- `JSONConfig`, `JSONSerializer`: added `includeNulls` switch
+- `JSONConfig`, `JSONSerializer`: added annotations to specify that null properties are to be included
+### Added
+- `JSONIncludeIfNull`: Annotation class
+- `JSONIncludeAllProperties`: Annotation class
+
 ## [3.0] - 2020-01-27
 ### Added
 - `JSONTypeRef`: implementation of the [TypeReference](https://gafter.blogspot.com/2006/12/super-type-tokens.html)